% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL_PRS.R
\name{block_calculation_pv_es}
\alias{block_calculation_pv_es}
\title{Update betas in each LD block for a list of learning rates using block-wise early stopping}
\usage{
block_calculation_pv_es(
  cor,
  num,
  nsnp,
  temp.file,
  lr_list,
  maxiter,
  sum_stats_target_val,
  patience,
  trace,
  BedFileReader
)
}
\arguments{
\item{cor}{Source summary statistics to update for each LD block}

\item{num}{LD block number}

\item{nsnp}{A number of total SNPs used in analysis}

\item{temp.file}{Output path to save intermediate results}

\item{lr_list}{A list of learning rates}

\item{maxiter}{A maximum number of iterations}

\item{sum_stats_target_val}{A matrix of target summary statistics for validation}

\item{patience}{A number of times to tolerate decrease of validation pseudo-R}

\item{trace}{Logical; if TRUE, print stopped iteration}

\item{BedFileReader}{A BedFileReader object to read reference Bedfile}
}
\value{
Information of updated betas (geno_info2), a list of numerators (beta_rho) and a matrix of PRS to calculate denominator (beta_g) for validation pseudo-R with all learning rates
}
\description{
Update betas in each LD block for a list of learning rates using block-wise early stopping
}
