% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p2cor.R
\name{p2cor}
\alias{p2cor}
\title{Function to convert p-values to correlation via the t-statistic}
\usage{
p2cor(p, n, sign = rep(1, length(p)), min.n = max(n, 30)/10)
}
\arguments{
\item{p}{Vector of p-values}

\item{n}{Sample size}

\item{sign}{A vector giving the sign of the correlations (e.g. the log odds ratios)}

\item{min.n}{The minimum sample size to be considered a valid p-value}
}
\value{
A vector of correlations
}
\description{
Function to convert p-values to correlation via the t-statistic
}
\references{
Mak, T. S. H., Porsch, R. M., Choi, S. W., Zhou, X., & Sham, P. C. (2017).
Polygenic scores via penalized regression on summary statistics.
Genetic Epidemiology, 41(6), 469–480. \doi{10.1002/gepi.22050}

See also the original lassosum package: \url{https://github.com/tshmak/lassosum}
}
