% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL_PRS.R
\name{PRStr_calculation_pv_es}
\alias{PRStr_calculation_pv_es}
\title{Update betas for all LD blocks for a list of learning rates using block-wise early stopping}
\usage{
PRStr_calculation_pv_es(
  sum_stats_target_train,
  ref_file,
  LDblocks,
  num_cores = 1,
  temp.file,
  lr_list,
  maxiter,
  sum_stats_target_val,
  patience,
  trace
)
}
\arguments{
\item{sum_stats_target_train}{Source summary statistics to update}

\item{ref_file}{Prefix of PLINK file of the target reference panel}

\item{LDblocks}{Character string specifying the LD blocks, defaults to "EUR.hg19".}

\item{num_cores}{A number of cores to be used in parallel computation for training, defaults to 1.}

\item{temp.file}{Output path to save intermediate results}

\item{lr_list}{A list of learning rates}

\item{maxiter}{A maximum number of iterations}

\item{sum_stats_target_val}{A matrix of target summary statistics for validation}

\item{patience}{A number of times to tolerate decrease of validation pseudo-R}

\item{trace}{Logical; if TRUE, print stopped iteration}
}
\value{
Information of updated betas (beta.byL), a list of numerators (betaRho.byL) and a matrix of PRS to calculate denominator (betaG.byL) for validation pseudo-R with all learning rates for all LD blocks
}
\description{
Update betas for all LD blocks for a list of learning rates using block-wise early stopping
}
