% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL_PRS.R
\name{pseudo_split}
\alias{pseudo_split}
\title{Pseudo-split target summary statistics into train and validation set}
\usage{
pseudo_split(target_sumstats, subprop, ref_file_ps, tempfile, random_seed = 42)
}
\arguments{
\item{target_sumstats}{Target summary statistics to split}

\item{subprop}{Proportion of training samples from full samples}

\item{ref_file_ps}{Prefix of PLINK file of the reference panel data in the target population}

\item{tempfile}{Output path to save splitted summary statistics}

\item{random_seed}{A random seed number to randomly (pseudo) split target summary statistics}
}
\value{
Splitted summary statistics for train and validation sammples
}
\description{
Pseudo-split target summary statistics into train and validation set
}
