% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL_PRS.R
\docType{package}
\name{PTL.PRS-package}
\alias{PTL.PRS}
\alias{PTL.PRS-package}
\title{PTL.PRS: Transfer learning of Polygenic Risk Score with Pseudo-validation}
\description{
Use transfer learning to fine-tune the baseline PRS model trained with GWAS summary statistics from a larger sample size source group to a smaller target group, using summary statistics only.
}
\author{
\strong{Maintainer}: Bokeum Cho \email{bokeum1810@snu.ac.kr}

}
