% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PTL_PRS_test}
\alias{PTL_PRS_test}
\title{Test performance of PTL-PRS}
\usage{
PTL_PRS_test(
  plink_file,
  by_chr,
  ped_test_file,
  outfile,
  Ytype,
  Covar_name,
  Y_name
)
}
\description{
Test performance of PTL-PRS
}
