% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PTL_PRS_test_pseudo}
\alias{PTL_PRS_test_pseudo}
\title{Test performance of PTL-PRS using pseudo test summary}
\usage{
PTL_PRS_test_pseudo(
  best.beta,
  best.param,
  sum_stats_target_test,
  sum_stats,
  ref_file,
  chunk_size = dim(best.beta)[1]
)
}
\arguments{
\item{best.beta}{Data.frame / matrix with SNP plus at least 5 cols}

\item{best.param}{learning rate}

\item{sum_stats_target_test}{data.frame of test summary stats}

\item{sum_stats}{data.frame of reference summary stats}

\item{ref_file}{prefix of PLINK .fam/.bim/.bed files}

\item{chunk_size}{Optional; number of SNPs to process per iteration}
}
\description{
Test performance of PTL-PRS using pseudo test summary
}
