% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PTL_PRS_test_pseudo_list}
\alias{PTL_PRS_test_pseudo_list}
\title{Test performance of PTL-PRS using pseudo test summary}
\usage{
PTL_PRS_test_pseudo_list(beta.list, sum_stats_target_test, sum_stats, ref_file)
}
\arguments{
\item{beta.list}{A list of beta coefficient vectors trained with all learning rates}

\item{sum_stats_target_test}{Target summary statistics used for testing}

\item{sum_stats}{Reference summary statistics (used in training)}

\item{ref_file}{Path to the reference genotype PLINK file}
}
\description{
Test performance of PTL-PRS using pseudo test summary
}
