% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL_PRS.R
\name{PTL_PRS_train}
\alias{PTL_PRS_train}
\title{Run PTL-PRS}
\usage{
PTL_PRS_train(
  ref_file,
  sum_stats_file,
  target_sumstats_file,
  subprop,
  ref_file_ps,
  LDblocks = "EUR.hg19",
  outfile,
  num_cores = 1,
  target_sumstats_train_file = NULL,
  target_sumstats_val_file = NULL,
  ps,
  pseudo_test,
  lr_list,
  iter,
  patience,
  trace = FALSE,
  random_seed = 42,
  random_seed_test = 70
)
}
\arguments{
\item{ref_file}{Prefix of PLINK file of the reference panel data in the target population.}

\item{sum_stats_file}{Location path of source PRS}

\item{target_sumstats_file}{Location path of target summary statistics}

\item{subprop}{Proportion of training samples from full samples}

\item{ref_file_ps}{Prefix of PLINK file of the reference panel data in the target population}

\item{LDblocks}{Character string specifying the LD blocks, defaults to "EUR.hg19".}

\item{outfile}{Output path to save training results}

\item{num_cores}{A number of cores to be used in parallel computation for training, defaults to 1.}

\item{target_sumstats_train_file}{Optional; Location path of target summary statistics for train samples. Defaults to NULL.}

\item{target_sumstats_val_file}{Optional; Location path of target summary statistics for validation samples. Defaults to NULL.}

\item{ps}{Logical; if TRUE, pseudo-split target summary statistics into train and validation.}

\item{pseudo_test}{Logical; if True, pseudo-split target summary statistics into train, validation and test.}

\item{lr_list}{A list of learning rates}

\item{iter}{A maximum number of iterations}

\item{patience}{A number of times to tolerate decrease of validation pseudo-R}

\item{trace}{Logical; if TRUE, print stopped iteration}

\item{random_seed}{A random seed number to split pseudo summary statistics for train and validation}

\item{random_seed_test}{A random seed number to split pseudo summary statistics for test and train+validation}
}
\value{
Returns a list containing the following elements:
\itemize{
\item \code{best.param}: The best learning rate determined by the validation.
\item \code{best.beta}: A list of the baseline and best beta coefficients found during validation. (SNP, CHR, A1, Beta2, best.beta)
\item \code{R2.list}: A list of R-squared values computed during the validation step.
}
}
\description{
Run PTL-PRS
}
