% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{readSomeSnp}
\alias{readSomeSnp}
\title{Read genotypes of provided SNPs and samples}
\usage{
readSomeSnp(snpList, sampleList = NULL, BedFileReader)
}
\description{
Read genotypes of provided SNPs and samples
}
