% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{TL_PRS_test}
\alias{TL_PRS_test}
\title{Test performance of PTL-PRS}
\usage{
TL_PRS_test(
  plink_file,
  ped_test_file,
  best.beta,
  best.param,
  Ytype,
  Covar_name,
  Y_name
)
}
\description{
Test performance of PTL-PRS
}
