% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PTL_PRS.R
\name{updateCoeff}
\alias{updateCoeff}
\title{Update betas in each LD block for each learning rate, using block-wise early stopping}
\usage{
updateCoeff(
  beta.all,
  GG2,
  geno_ref,
  learningrate,
  maxiter,
  betaSD,
  sum_stats_target_val_block,
  patience,
  trace
)
}
\arguments{
\item{beta.all}{A list of two elements: baseline betas to update for each LD block (betatemp), and their gradients (gy-GG2\%*\%betatemp)}

\item{GG2}{A covariance matrix of reference genotypes (p'*p')}

\item{geno_ref}{A genotype matrix of reference panel (n*p')}

\item{learningrate}{Learning rate}

\item{maxiter}{A maximum number of iterations}

\item{betaSD}{Standard deviation of each SNP's beta}

\item{sum_stats_target_val_block}{A matrix of target summary statistics for each LD block}

\item{trace}{Logical; if TRUE, print stopped iteration}
}
\value{
Updated betas (betatemp), numerator (R_num) and a list of PRS to calculate denominator (PRS) for validation pseudo-R
}
\description{
Update betas in each LD block for each learning rate, using block-wise early stopping
}
