% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicFunctions.R
\name{PTL_PRS_test_pseudo}
\alias{PTL_PRS_test_pseudo}
\title{Test performance of PTL-PRS using pseudo test summary}
\usage{
PTL_PRS_test_pseudo(
  best.beta,
  best.param,
  sum_stats_target_test,
  sum_stats,
  ref_file
)
}
\arguments{
\item{best.beta}{A list of beta coefficient vectors, containing baseline and the best set}

\item{best.param}{A learning rate used with the best beta}

\item{sum_stats_target_test}{Target summary statistics used for testing}

\item{sum_stats}{Reference summary statistics (used in training)}

\item{ref_file}{Path to the reference genotype PLINK file}
}
\description{
Test performance of PTL-PRS using pseudo test summary
}
